% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_value}
\alias{get_value}
\title{Get a value from a TOML document.}
\usage{
get_value(toml_doc, key_path)
}
\arguments{
\item{toml_doc}{A TOML document.}

\item{key_path}{A path to the value.}
}
\description{
Get a value from a TOML document.
}
